requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.43.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
asyncer==0.0.8
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
cohere==5.13.11
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
dataclasses-json==0.6.7
datasets==2.14.7
decorator==5.1.1
Deprecated==1.2.14
devtools==0.12.2
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
dominate==2.9.1
dspy==2.5.40
dspy-ai==2.5.40
email_validator==2.1.1
et-xmlfile==1.1.0
eval_type_backport==0.2.2
executing==2.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.10.0
filelock==3.14.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.1
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SocketIO==5.4.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2023.10.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
griffe==1.5.4
groq==0.15.0
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
joblib==1.3.2
json_repair==0.30.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==29.0.0
langchain==0.3.14
langchain-chroma==0.2.0
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-redis==0.1.2
langchain-text-splitters==0.3.5
langfuse==2.57.5
langsmith==0.2.10
litellm==1.51.0
logfire==3.2.0
logfire-api==3.1.1
magicattr==0.1.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.3.0
ml-dtypes==0.4.1
mmh3==4.1.0
mongoengine==0.28.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.64.0
openpyxl==3.1.5
# opentelemetry-api==1.29.0
# opentelemetry-exporter-otlp-proto-common==1.29.0
# opentelemetry-exporter-otlp-proto-grpc==1.24.0
# opentelemetry-exporter-otlp-proto-http==1.29.0
# opentelemetry-instrumentation==0.45b0
# opentelemetry-instrumentation-asgi==0.45b0
# opentelemetry-instrumentation-fastapi==0.45b0
# opentelemetry-proto==1.29.0
# opentelemetry-sdk==1.29.0
# opentelemetry-semantic-conventions==0.50b0
# opentelemetry-util-http==0.45b0
optuna==3.6.1
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
playwright==1.50.0
pluggy==1.5.0
posthog==3.5.0
prompt_toolkit==3.0.48
protobuf==5.29.3
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.10.6
pydantic-ai==0.0.25
pydantic-ai-slim==0.0.25
pydantic-graph==0.0.25
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.18.0
pymongo==4.7.2
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypandoc==1.13
pypdf==4.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pysqlite3==0.5.3
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-slugify==8.0.4
python-socketio==5.11.4
python-ulid==2.7.0
pytz==2024.1
PyYAML==6.0.1
redis==5.2.1
redisvl==0.3.8
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.21.0
rsa==4.9
setuptools==70.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
structlog==24.4.0
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
text-unidecode==1.3
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
URLObject==2.4.3
uvicorn==0.30.0
uvloop==0.19.0
visitor==0.1.3
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.0
